weeeeeeeeeee

+++ src/main.h
@@ -1,18 +1,16 @@
 #ifndef DMM_MAIN_H
 #define DMM_MAIN_H
 
-// ASSUMPTIONS: Requires stdint.h with uint8_t and stddef.h with size_t.
+// REQUIREMENTS: Requires size_t (stddef.h).
 
-#include <stdint.h>
 #include <stddef.h>
-
 
 typedef void *(DMM_MallocFn)(size_t size);
 typedef void (DMM_FreeFn)(void *ptr);
 
 typedef struct dmm_malloc_header_s {
     size_t size;
-    uint8_t used;
+    size_t used; // A bit space-inefficient, but means we only require one type.
     void *data;
 } DMM_MallocHeader;
